chore: rename x/dibc to x/vibc
