+++ golang/cosmos/app/app.go
@@ -97,8 +97,8 @@ import (
 
 	gaiaappparams "github.com/Agoric/agoric-sdk/golang/cosmos/app/params"
 	"github.com/Agoric/agoric-sdk/golang/cosmos/vm"
-	"github.com/Agoric/agoric-sdk/golang/cosmos/x/dibc"
 	"github.com/Agoric/agoric-sdk/golang/cosmos/x/swingset"
+	"github.com/Agoric/agoric-sdk/golang/cosmos/x/vibc"
 	"github.com/Agoric/agoric-sdk/golang/cosmos/x/vpurse"
 
 	// unnamed import of statik for swagger UI support
@@ -130,7 +130,7 @@ var (
 		slashing.AppModuleBasic{},
 		ibc.AppModuleBasic{},
 		swingset.AppModuleBasic{},
-		dibc.AppModuleBasic{},
+		vibc.AppModuleBasic{},
 		vpurse.AppModuleBasic{},
 		feegrantmodule.AppModuleBasic{},
 		upgrade.AppModuleBasic{},
@@ -197,13 +197,13 @@ type GaiaApp struct { // nolint: golint
 	TransferKeeper   ibctransferkeeper.Keeper
 
 	SwingSetKeeper swingset.Keeper
-	DibcKeeper     dibc.Keeper
-	VpurseKeeper   vpurse.Keeper
+	VibcKeeper     vibc.Keeper
+	VbankKeeper    vpurse.Keeper
 
 	// make scoped keepers public for test purposes
 	ScopedIBCKeeper      capabilitykeeper.ScopedKeeper
 	ScopedTransferKeeper capabilitykeeper.ScopedKeeper
-	ScopedDibcKeeper     capabilitykeeper.ScopedKeeper
+	ScopedVibcKeeper     capabilitykeeper.ScopedKeeper
 
 	FeeGrantKeeper feegrantkeeper.Keeper
 	// the module manager
@@ -260,7 +260,7 @@ func NewAgoricApp(
 		minttypes.StoreKey, distrtypes.StoreKey, slashingtypes.StoreKey,
 		govtypes.StoreKey, paramstypes.StoreKey, ibchost.StoreKey, upgradetypes.StoreKey,
 		evidencetypes.StoreKey, feegrant.StoreKey, ibctransfertypes.StoreKey,
-		swingset.StoreKey, dibc.StoreKey, vpurse.StoreKey,
+		swingset.StoreKey, vibc.StoreKey, vpurse.StoreKey,
 		capabilitytypes.StoreKey,
 		authzkeeper.StoreKey,
 	)
@@ -286,7 +286,7 @@ func NewAgoricApp(
 	app.CapabilityKeeper = capabilitykeeper.NewKeeper(appCodec, keys[capabilitytypes.StoreKey], memKeys[capabilitytypes.MemStoreKey])
 	scopedIBCKeeper := app.CapabilityKeeper.ScopeToModule(ibchost.ModuleName)
 	scopedTransferKeeper := app.CapabilityKeeper.ScopeToModule(ibctransfertypes.ModuleName)
-	scopedDibcKeeper := app.CapabilityKeeper.ScopeToModule(dibc.ModuleName)
+	scopedVibcKeeper := app.CapabilityKeeper.ScopeToModule(vibc.ModuleName)
 
 	// add keepers
 	app.AccountKeeper = authkeeper.NewAccountKeeper(
@@ -369,31 +369,31 @@ func NewAgoricApp(
 	)
 	vm.RegisterPortHandler("storage", swingset.NewStorageHandler(app.SwingSetKeeper))
 
-	app.DibcKeeper = dibc.NewKeeper(
-		appCodec, keys[dibc.StoreKey],
+	app.VibcKeeper = vibc.NewKeeper(
+		appCodec, keys[vibc.StoreKey],
 		app.IBCKeeper.ChannelKeeper, &app.IBCKeeper.PortKeeper,
 		app.BankKeeper,
-		scopedDibcKeeper,
+		scopedVibcKeeper,
 		callToController,
 	)
 
-	dibcModule := dibc.NewAppModule(app.DibcKeeper)
-	app.ibcPort = vm.RegisterPortHandler("dibc", dibc.NewPortHandler(dibcModule, app.DibcKeeper))
+	vibcModule := vibc.NewAppModule(app.VibcKeeper)
+	app.ibcPort = vm.RegisterPortHandler("vibc", vibc.NewPortHandler(vibcModule, app.VibcKeeper))
 
 	// Create static IBC router, add transfer route, then set and seal it
 	// FIXME: Don't be confused by the name!  The port router maps *module names* (not PortIDs) to modules.
 	ibcRouter := porttypes.NewRouter()
 	ibcRouter.AddRoute(ibctransfertypes.ModuleName, transferModule)
-	ibcRouter.AddRoute(dibc.ModuleName, dibcModule)
+	ibcRouter.AddRoute(vibc.ModuleName, vibcModule)
 	app.IBCKeeper.SetRouter(ibcRouter)
 
-	app.VpurseKeeper = vpurse.NewKeeper(
+	app.VbankKeeper = vpurse.NewKeeper(
 		appCodec, keys[vpurse.StoreKey],
 		app.BankKeeper, authtypes.FeeCollectorName,
 		callToController,
 	)
-	vpurseModule := vpurse.NewAppModule(app.VpurseKeeper)
-	app.vpursePort = vm.RegisterPortHandler("bank", vpurse.NewPortHandler(vpurseModule, app.VpurseKeeper))
+	vbankModule := vpurse.NewAppModule(app.VbankKeeper)
+	app.vpursePort = vm.RegisterPortHandler("bank", vpurse.NewPortHandler(vbankModule, app.VbankKeeper))
 
 	// create evidence keeper with router
 	evidenceKeeper := evidencekeeper.NewKeeper(
@@ -436,8 +436,8 @@ func NewAgoricApp(
 		ibc.NewAppModule(app.IBCKeeper),
 		params.NewAppModule(app.ParamsKeeper),
 		swingset.NewAppModule(app.SwingSetKeeper),
-		dibcModule,
-		vpurseModule,
+		vibcModule,
+		vbankModule,
 		transferModule,
 		authzmodule.NewAppModule(appCodec, app.AuthzKeeper, app.AccountKeeper, app.BankKeeper, app.interfaceRegistry),
 	)
@@ -537,7 +537,7 @@ func NewAgoricApp(
 		app.CapabilityKeeper.InitializeAndSeal(ctx)
 	}
 	app.ScopedIBCKeeper = scopedIBCKeeper
-	app.ScopedDibcKeeper = scopedDibcKeeper
+	app.ScopedVibcKeeper = scopedVibcKeeper
 	app.ScopedTransferKeeper = scopedTransferKeeper
 
 	return app
