+++ golang/cosmos/x/vibc/types/key.go
@@ -2,7 +2,7 @@ package types
 
 const (
 	// module name
-	ModuleName = "dibc"
+	ModuleName = "vibc"
 
 	// StoreKey to be used when creating the KVStore
 	StoreKey = ModuleName
